requirements.txt
asgiref==3.7.2
blinker==1.6.2
click==8.1.3
Django==3.2.19
django-ratelimit==4.0.0
djangorestframework==3.14.0
djangorestframework-gis==1.0
factory-boy==3.2.1
Faker==18.11.1
Flask==2.3.2
GDAL==3.4.3
geos==0.2.3
gunicorn==20.1.0
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.2
MarkupSafe==2.1.3
Pillow==9.5.0
psycopg2-binary==2.9.6
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.3
Werkzeug==2.3.6
zipp==3.15.0